+++ apps/builder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @substrat-run/builder
 
+## 0.5.43
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/kernel@0.101.0
+  - @substrat-run/adapter-cloudflare@0.101.0
+  - @substrat-run/engine-metering@0.5.7
+  - @substrat-run/model-providers@0.4.4
+
 ## 0.5.42
 
 ### Patch Changes