+++ apps/router/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @substrat-run/router
 
+## 0.2.41
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/adapter-cloudflare@0.101.0
+
 ## 0.2.40
 
 ### Patch Changes