+++ demos/callout/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @substrat-run/demo-fsm
 
+## 0.3.19
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+- Updated dependencies [8406843]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/kernel@0.101.0
+  - @substrat-run/adapter-sqlite@0.101.0
+  - @substrat-run/adapter-cloudflare@0.101.0
+  - @substrat-run/control-plane-api@0.101.0
+  - @substrat-run/engine-workorder@0.11.0
+  - @substrat-run/engine-invoicing@0.9.17
+  - @substrat-run/engine-protocol@0.12.6
+  - @substrat-run/dev-issuer@0.1.16
+  - @substrat-run/vertical-auth@0.12.4
+
 ## 0.3.18
 
 ### Patch Changes