+++ demos/manyfold/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @substrat-run/demo-manyfold
 
+## 0.7.18
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/kernel@0.101.0
+  - @substrat-run/adapter-sqlite@0.101.0
+  - @substrat-run/adapter-cloudflare@0.101.0
+  - @substrat-run/dev-issuer@0.1.16
+  - @substrat-run/vertical-auth@0.12.4
+  - @substrat-run/vertical-host@0.101.0
+
 ## 0.7.17
 
 ### Patch Changes