+++ demos/shop/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @substrat-run/demo-shop
 
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/kernel@0.101.0
+  - @substrat-run/adapter-sqlite@0.101.0
+  - @substrat-run/engine-invoicing@0.9.17
+  - @substrat-run/dev-issuer@0.1.16
+  - @substrat-run/vertical-host@0.101.0
+
 ## 0.2.12
 
 ### Patch Changes