+++ packages/vertical-host/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @substrat-run/vertical-host
 
+## 0.101.0
+
+### Patch Changes
+
+- Updated dependencies [b61c4d5]
+- Updated dependencies [306b893]
+  - @substrat-run/contracts@0.101.0
+  - @substrat-run/kernel@0.101.0
+  - @substrat-run/model-providers@0.4.4
+
 ## 0.100.0
 
 ### Patch Changes